requirements.txt
fastapi
uvicorn
scikit-learn==1.2.2
numpy
scipy
matplotlib
pandas


httptools

torch
torchvision
pybase64
pillow

